ums/requirements.txt
#ums/requirements.txt

mysql-connector-python
flask